# docs/swagger.yaml
basePath: /api
definitions:
  auth.CredentialsData:
    properties:
      access_token:
        $ref: '#/definitions/shared.TokenData'
      refresh_token:
        $ref: '#/definitions/shared.TokenData'
    type: object
  auth.ForgotPasswordRequest:
    properties:
      email:
        example: johndoe@example.com
        type: string
    required:
    - email
    type: object
  auth.ForgotPasswordResponse:
    properties:
      reset_token:
        $ref: '#/definitions/shared.TokenData'
      to_email:
        type: string
    type: object
  auth.LoginRequest:
    properties:
      email:
        example: johndoe@example.com
        type: string
      password:
        example: password123
        type: string
    required:
    - email
    - password
    type: object
  auth.LoginResponse:
    properties:
      credentials:
        $ref: '#/definitions/auth.CredentialsData'
      user:
        $ref: '#/definitions/user.UserResponse'
    type: object
  auth.RegisterRequest:
    properties:
      email:
        example: johndoe@example.com
        type: string
      name:
        example: John Doe
        type: string
      password:
        example: password123
        minLength: 8
        type: string
    required:
    - email
    - name
    - password
    type: object
  auth.RegisterResponse:
    properties:
      user:
        $ref: '#/definitions/user.UserResponse'
    type: object
  auth.ResetPasswordRequest:
    properties:
      new_password:
        example: newpassword123
        minLength: 8
        type: string
      token:
        type: string
    required:
    - new_password
    - token
    type: object
  shared.TokenData:
    properties:
      expired_at:
        type: string
      token:
        type: string
    type: object
  user.ProfileResponse:
    properties:
      date_birth:
        example: "1990-01-01"
        type: string
      first_name:
        example: John
        type: string
      gender:
        example: Male
        type: string
      last_name:
        example: Doe
        type: string
      photo:
        example: https://example.com/photo.jpg
        type: string
    type: object
  user.UserResponse:
    properties:
      email:
        example: johndoe@example.com
        type: string
      id:
        example: "1"
        type: string
      profile:
        $ref: '#/definitions/user.ProfileResponse'
      username:
        example: johndoe
        type: string
    type: object
info:
  contact: {}
  description: Pradnya Backend API
  title: Pradnya API
  version: "1.0"
paths:
  /auth/forgot-password:
    post:
      consumes:
      - application/json
      description: Send reset password email
      parameters:
      - description: Forgot Password Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.ForgotPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.ForgotPasswordResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: Forgot Password
      tags:
      - Auth
  /auth/login:
    post:
      consumes:
      - application/json
      description: Verify User Cred
      parameters:
      - description: Login Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.LoginResponse'
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
      summary: Login user
      tags:
      - Auth
  /auth/me:
    get:
      consumes:
      - application/json
      description: User
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.UserResponse'
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: Me user
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Create new account
      parameters:
      - description: Register Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/auth.RegisterResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Register user
      tags:
      - Auth
  /auth/reset-password:
    post:
      consumes:
      - application/json
      description: Reset user password
      parameters:
      - description: Reset Password Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.ResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: Reset Password
      tags:
      - Auth
swagger: "2.0"